#118 refactor: Get user api 추가
