+++ src/api/user/queries/usePasswordEditMutation.ts
@@ -1,6 +1,6 @@
 import { useMutation } from "@tanstack/react-query";
 import { putNewPassword } from "..";
-import { settingsKeys } from "./queryKeys";
+import { userKeys } from "./queryKeys";
 
 type Props = {
   onSuccess: () => void;
@@ -8,7 +8,7 @@ type Props = {
 
 export default function usePasswordEditMutation({ onSuccess }: Props) {
   return useMutation({
-    mutationKey: settingsKeys.passwordEdit().queryKey,
+    mutationKey: userKeys.passwordEdit().queryKey,
     mutationFn: putNewPassword,
     onSuccess,
     meta: {
