+++ src/api/user/queries/useProfileDetailsMutation.ts
@@ -2,13 +2,13 @@ import { UserContext } from "@context/UserContext";
 import { useMutation } from "@tanstack/react-query";
 import { useContext } from "react";
 import { putProfileDetails } from "..";
-import { settingsKeys } from "./queryKeys";
+import { userKeys } from "./queryKeys";
 
 export default function useProfileDetailsMutation() {
   const { onEditProfileDetails } = useContext(UserContext);
 
   return useMutation({
-    mutationKey: settingsKeys.profileEdit().queryKey,
+    mutationKey: userKeys.profileEdit().queryKey,
     mutationFn: putProfileDetails,
     onSuccess: (res) => {
       onEditProfileDetails(res.data.user);
