+++ src/api/user/types.ts
@@ -1,0 +1,12 @@
+export type User = {
+  id: number;
+  nickname: string;
+  email: string;
+  profileUrl: string;
+  notificationPreferences: {
+    browserNotify: boolean;
+    targetGainNotify: boolean;
+    maxLossNotify: boolean;
+    targetPriceNotify: boolean;
+  };
+};
